requirements.txt
altgraph==0.17.4
dearpygui==1.10.1
packaging==23.2
pefile==2023.2.7
pycryptodomex==3.20.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyperclip==1.8.2
pywin32-ctypes==0.2.2
